pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "videoseek"
version = "0.1.0"
description = "VideoSeek CLI and agent package"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "decord",
  "litellm",
  "numpy",
  "pillow",
  "pyyaml",
  "yt-dlp",
  "openai-whisper",
]

[project.scripts]
"videoseek-cli" = "videoseek.cli:main"

[tool.setuptools]
packages = { find = { include = ["videoseek*", "config*"] } }
